update: code correction about varaibles type from int to double

+++ cfib.pyx
@@ -1,9 +1,9 @@
 # cython: language_level=3
 cimport cython
 
-cpdef int cython_fib(int n):
+cpdef double cython_fib(int n):
     cdef int i
-    cdef int a = 0, b = 1
+    cdef double a = 0.0, b = 1.0
     for i in range(n):
         a, b = a + b, a
     return a
